+++ AoE2ScenarioParser/settings.py
@@ -20,6 +20,8 @@ ALLOW_OVERWRITING_SOURCE = False
 """Disable the error being raised when overwriting source scenario"""
 ALLOW_DIRTY_RETRIEVER_OVERWRITE = False
 """If it is allowed to overwrite a retriever that is dirty (it has been changed manually)"""
+SHOW_VARIANT_WARNINGS = True
+"""If warnings about incorrect variants should be shown or not"""
 
 # Charset settings
 MAIN_CHARSET = "utf-8"